# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: OrigCMS API
    description: OrigCMS is a modern Content Management System for video platforms.
    contact:
        name: GodCong
        url: https://github.com/origadmin
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://origadmin/application/origcms/blob/master/LICENSE
    version: Version from annotation
servers:
    - url: http://localhost:18080
    - url: https://localhost:18080
paths:
...
    /api/v1/admin/stats/medias:
        get:
            tags:
                - AdminService
            operationId: AdminService_GetMediaStats
            parameters:
                - name: period
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaStatsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/admin/stats/users:
        get:
            tags:
                - AdminService
            operationId: AdminService_GetUserStats
            parameters:
                - name: period
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetUserStatsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias:
        get:
            tags:
                - MediaService
            description: ListMedias returns a list of medias.
            operationId: MediaService_ListMedias
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page_token
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: category_id
                  in: query
                  schema:
                    type: string
                - name: user_id
                  in: query
                  schema:
                    type: string
                - name: order_by
                  in: query
                  schema:
                    type: string
                - name: descending
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.ListMediasResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - MediaService
            description: CreateMedia creates a new media.
            operationId: MediaService_CreateMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.CreateMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.CreateMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/upload:
        post:
            tags:
                - MediaService
            description: UploadMedia uploads a media file.
            operationId: MediaService_UploadMedia
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.UploadMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.UploadMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{id}:
        get:
            tags:
                - MediaService
            description: GetMedia returns a media by ID.
            operationId: MediaService_GetMedia
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - MediaService
            description: DeleteMedia deletes a media.
            operationId: MediaService_DeleteMedia
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.DeleteMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{id}/download:
        get:
            tags:
                - MediaService
            description: GetMediaDownload returns the download URL for a media.
            operationId: MediaService_GetMediaDownload
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaDownloadResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{id}/stream:
        get:
            tags:
                - MediaService
            description: GetMediaStream returns the streaming URL for a media.
            operationId: MediaService_GetMediaStream
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaStreamResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /api/v1/medias/{id}/thumbnail:
        get:
            tags:
                - MediaService
            description: GetMediaThumbnail returns the thumbnail URL for a media.
            operationId: MediaService_GetMediaThumbnail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaThumbnailResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/{media.id}:
        put:
            tags:
                - MediaService
            description: UpdateMedia updates an existing media.
            operationId: MediaService_UpdateMedia
            parameters:
                - name: media.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.media.UpdateMediaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.UpdateMediaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /api/v1/medias/{media_id}/metadata:
        get:
            tags:
                - MediaService
            description: GetMediaMetadata returns metadata for a media.
            operationId: MediaService_GetMediaMetadata
            parameters:
                - name: media_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.media.GetMediaMetadataResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
components:
    schemas:
...
        api.v1.services.media.CreateMediaRequest:
            type: object
            properties:
                media:
                    $ref: '#/components/schemas/api.v1.services.types.Media'
...
        api.v1.services.media.DeleteMediaResponse:
            type: object
            properties: {}
...
        api.v1.services.media.GetMediaStatsResponse:
            type: object
            properties:
                total_uploads:
                    type: integer
                    format: int32
                total_views:
                    type: integer
                    format: int32
                daily_stats:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.StatPoint'
...
        api.v1.services.media.GetUserStatsResponse:
            type: object
            properties:
                total_users:
                    type: integer
                    format: int32
                new_users:
                    type: integer
                    format: int32
                active_users:
                    type: integer
                    format: int32
                daily_stats:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.types.StatPoint'
...
    securitySchemes:
        Basic:
            type: http
            scheme: basic
        Bearer:
            type: apiKey
            name: Authorization
            in: header
...